Return something more reasonable for n.sv=0
In particular, when n.sv is zero, the returned object now has a
zero-column SV matrix, which means that one could write code that does
cbind(design, svobj$sv) without special-casing zero.

Additionally, when n.sv is explicitly set to zero by the user (as
opposed to automatically determined to be zero), a warning is issued.